stack-o11y/requirements.txt
opentelemetry-sdk==1.27.0 
opentelemetry-exporter-otlp==1.27.0